+++ swift_book_pdf/book.py
@@ -32,7 +32,7 @@ class Book:
         self.converter = LaTeXConverter(config)
 
     def process_files_in_order(self, latex_file_path: str):
-        latex = generate_preamble(self.config.rendering_mode)
+        latex = generate_preamble(self.config.doc_config)
         # TODO: Use the version to generate a cover page
         toc_latex, _ = self.toc.generate_toc_latex(converter=self.converter)
         latex += toc_latex + "\n"
@@ -54,9 +54,9 @@ class Book:
     def process(self):
         latex_file_path = os.path.join(self.config.temp_dir, "inner_content.tex")
         self.process_files_in_order(latex_file_path)
-        logger.info(f"Creating PDF in {self.config.rendering_mode.value} mode...")
+        logger.info(f"Creating PDF in {self.config.doc_config.mode.value} mode...")
         converter = PDFConverter(self.config)
-        for _ in trange(self.config.typesets, leave=False):
+        for _ in trange(self.config.doc_config.typesets, leave=False):
             converter.convert_to_pdf(latex_file_path)
 
         temp_pdf_path = os.path.join(self.config.temp_dir, "inner_content.pdf")